+++ .beads/issues.jsonl
@@ -3,6 +3,7 @@
 {"_type":"issue","id":"blog-aei","title":"Add PostHog tracking alongside Umami","description":"Add PostHog client-side pageview tracking through the Vercel PostHog integration while keeping existing Umami and Vercel Analytics active.","acceptance_criteria":"PostHog provider is wired with Vercel integration env vars; Umami script remains active; Vercel Analytics remains active; dashboard source remains Umami during warm-up.","status":"closed","priority":1,"issue_type":"task","owner":"hello@ruchern.dev","created_at":"2026-05-03T17:28:41Z","created_by":"Ru Chern Chong","updated_at":"2026-05-03T17:30:02Z","closed_at":"2026-05-03T17:30:02Z","close_reason":"completed","labels":["analytics","migration","posthog","umami"],"dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"blog-7ow","title":"Set up HeroUI Pro foundation","description":"Add and configure the HeroUI and HeroUI Pro foundation needed for the blog redesign. Use the Default preset as the target, preserve the correct CSS import order, and avoid activating Brutalism or Glass globally. This work should document how HeroUI Pro coexists with shadcn/ui while shadcn remains legacy during migration.","acceptance_criteria":"@heroui/react and @heroui-pro/react are available; CSS imports follow HeroUI Pro requirements; no global Brutalism or Glass theme is enabled; project documentation or implementation notes explain the temporary coexistence with shadcn/ui.","notes":"GitHub issue: https://github.com/ruchernchong/blog/issues/300","status":"open","priority":1,"issue_type":"task","owner":"hello@ruchern.dev","created_at":"2026-04-30T15:53:59Z","created_by":"Ru Chern Chong","updated_at":"2026-04-30T16:46:29Z","dependencies":[{"issue_id":"blog-7ow","depends_on_id":"blog-okc","type":"blocks","created_at":"2026-04-30T23:54:33Z","created_by":"Ru Chern Chong","metadata":"{}"}],"dependency_count":1,"dependent_count":2,"comment_count":0}
 {"_type":"issue","id":"blog-okc","title":"Create Blog DESIGN.md","description":"Define the public blog design direction using the google-labs-code DESIGN.md format. The document should use name: Blog, describe HeroUI Pro Default as the target design system, mark shadcn/ui as legacy during migration, and capture the warm editorial colour direction, typography, layout, components, and do/don't rules for the blog redesign.","acceptance_criteria":"DESIGN.md exists at the repo root; it follows the DESIGN.md alpha section order; it uses name: Blog; it identifies HeroUI Pro Default as the target; it treats shadcn/ui as legacy during migration; it includes blog-specific component guidance and do/don't rules.","notes":"GitHub issue: https://github.com/ruchernchong/blog/issues/299","status":"open","priority":1,"issue_type":"task","owner":"hello@ruchern.dev","created_at":"2026-04-30T15:53:47Z","created_by":"Ru Chern Chong","updated_at":"2026-04-30T16:46:22Z","dependency_count":0,"dependent_count":3,"comment_count":0}
+{"_type":"issue","id":"blog-k2j","title":"Standardise motion imports on RSC-safe motion/react-client","description":"Convert hero-section, featured-work, not-found, and dashboard views-by-page from 'use client' + motion/react to Server Components using motion/react-client, matching latest-posts.tsx. Removes unnecessary client boundaries.","status":"open","priority":2,"issue_type":"task","owner":"hello@ruchern.dev","created_at":"2026-05-30T20:06:11Z","created_by":"Ru Chern Chong","updated_at":"2026-05-30T20:06:11Z","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"blog-55p","title":"Complete Phase F: remove all shadcn remnants","description":"Migration left the shadcn foundation in place: token layer + @theme inline bridge in globals.css, cn/cva utilities, dead deps (recharts, react-resizable-panels, class-variance-authority, clsx, tailwind-merge, tw-animate-css), shadcn MCP/skill registration, and stale docs. Re-pin HeroUI tokens to coral/warm-paper, rename shadcn utility classes to HeroUI equivalents, strip deps/config/docs. Zero visual change. Plan: ~/.claude/plans/check-if-any-remnant-starry-owl.md","status":"closed","priority":2,"issue_type":"task","assignee":"Ru Chern Chong","owner":"hello@ruchern.dev","created_at":"2026-05-30T18:32:09Z","created_by":"Ru Chern Chong","updated_at":"2026-05-30T18:47:54Z","started_at":"2026-05-30T18:32:19Z","closed_at":"2026-05-30T18:47:54Z","close_reason":"Phase F complete: shadcn foundation removed (tokens re-pinned to HeroUI, cn from @heroui/react, utility classes migrated, deps/config/docs cleaned). recharts/react-resizable-panels/tailwind-merge kept as HeroUI Pro peer deps. typecheck/lint/build green; only pre-existing MCP-route test failures remain.","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"blog-a4z","title":"Phase F: Delete shadcn + unused deps","description":"Delete src/components/ui, components.json, shadcn/tailwind.css import; remove shadcn/@base-ui/react/recharts/react-resizable-panels deps (KEEP hugeicons). pnpm install + full build/test. Part of blog-zs8.","status":"closed","priority":2,"issue_type":"task","assignee":"Ru Chern Chong","owner":"hello@ruchern.dev","created_at":"2026-05-30T07:59:14Z","created_by":"Ru Chern Chong","updated_at":"2026-05-30T13:11:26Z","started_at":"2026-05-30T13:04:33Z","closed_at":"2026-05-30T13:11:26Z","close_reason":"shadcn fully removed: deleted src/components/ui (24 files), components.json, dead shadcn/tailwind.css import; removed shadcn/@base-ui/react/recharts/react-resizable-panels deps (kept hugeicons/cva/clsx/tailwind-merge/tailwind-variants). Build 65/65 pages, typecheck+lint clean. Migration complete.","dependencies":[{"issue_id":"blog-a4z","depends_on_id":"blog-zs8","type":"blocks","created_at":"2026-05-30T15:59:44Z","created_by":"Ru Chern Chong","metadata":"{}"}],"dependency_count":1,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"blog-95p","title":"Phase E: Migrate auth components to HeroUI","description":"login form + user-menu (button, dropdown, dialog, avatar). Part of blog-zs8.","status":"closed","priority":2,"issue_type":"task","owner":"hello@ruchern.dev","created_at":"2026-05-30T07:59:13Z","created_by":"Ru Chern Chong","updated_at":"2026-05-30T13:03:56Z","closed_at":"2026-05-30T13:03:56Z","close_reason":"Studio CMS + auth migrated to HeroUI: forms (TextField/Select/Checkbox), tables (AlertDialog/EmptyState/Chip), Modal dialogs, Pro Sidebar shell, content-editor Tabs/Resizable. No @/components/ui imports remain. Typecheck clean.","dependencies":[{"issue_id":"blog-95p","depends_on_id":"blog-zs8","type":"blocks","created_at":"2026-05-30T15:59:43Z","created_by":"Ru Chern Chong","metadata":"{}"}],"dependency_count":1,"dependent_count":0,"comment_count":0}
