+++ src/app/not-found.tsx
@@ -1,8 +1,6 @@
-"use client";
-
 import { cn } from "@heroui/react";
 import { buttonVariants } from "@heroui/styles";
-import { motion } from "motion/react";
+import * as motion from "motion/react-client";
 import Link from "next/link";
 import { BackgroundEffects } from "@/components/background-effects";
